pallet_people
Module pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

pub use __tt_extra_parts_7 as tt_extra_parts;

pub use __tt_default_parts_v2_7 as tt_default_parts_v2;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

EnsurePersonalAlias

Guard to ensure that the given origin is a person. The contextual alias of the person is provided on success.

EnsurePersonalAliasInContext

Guard to ensure that the given origin is a person. The alias of the person within the context provided as an argument is returned on success.

EnsurePersonalIdentity

Guard to ensure that the given origin is a person. The underlying identity of the person is provided on success.

EnsureRevisedPersonalAlias

Guard to ensure that the given origin is a person.

EnsureRevisedPersonalAliasInContext

Guard to ensure that the given origin is a person.

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Origin

Traits

Config

Configuration trait of this pallet.

Functions

ensure_personal_alias

Ensure that the origin o represents an extrinsic (i.e. transaction) from a personal alias. Returns Ok with the personal alias that signed the extrinsic or an Err otherwise.

ensure_personal_identity

Ensure that the origin o represents an extrinsic (i.e. transaction) from a personal identity. Returns Ok with the personal identity that signed the extrinsic or an Err otherwise.

ensure_revised_personal_alias

Ensure that the origin o represents an extrinsic (i.e. transaction) from a personal alias with revision information.

Type Aliases

AccountToAlias

Conversion of an account ID to a contextual alias.

AccountToPersonalId

Association of an account ID to a personal ID.

ActiveMembers

The number of people currently included in a ring.

AliasToAccount

Conversion of a contextual alias to an account ID.

Chunks

Paginated collection of static chunks used by the verifiable crypto.

CurrentRingIndex

Keeps track of the ring index currently being populated.

KeyMigrationQueue

A map of all the people who have declared their intent to migrate their keys and are waiting for the next mutation session.

Keys

The current individuals we recognise, but not necessarily yet included in a ring.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NextPersonalId

The next free and never reserved personal ID.

OnboardingQueue

Paginated collection of people public keys ready to be included in a ring.

OnboardingSize

Maximum number of people queued before onboarding to a ring.

PendingSuspensions

A map of all rings which currently have pending suspensions and need cleaning, along with their respective number of suspended keys which need to be removed.

People

The current individuals we recognise, but not necessarily yet included in a ring.

QueuePageIndices

Keeps track of the page indices of the head and tail of the onboarding queue.

ReservedPersonalId

Candidates’ reserved identities which we track.

RingBuildingPeopleLimit

Hint for the maximum number of people that can be included in a ring through a single root building call. If no value is set, then the onboarding size will be used instead.

RingKeys

Both the keys that are included in built rings and the keys that will be used in future rings.

RingKeysStatus

Stores the meta information for each ring, the number of keys and how many are actually included in the root.

RingsState

The state of the pallet regarding the actions that are currently allowed to be performed on all existing rings.

Root

The current individuals we recognise.