pallet_people::pallet
Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

AccountToAlias

Conversion of an account ID to a contextual alias.

AccountToPersonalId

Association of an account ID to a personal ID.

ActiveMembers

The number of people currently included in a ring.

AliasToAccount

Conversion of a contextual alias to an account ID.

Chunks

Paginated collection of static chunks used by the verifiable crypto.

CurrentRingIndex

Keeps track of the ring index currently being populated.

KeyMigrationQueue

A map of all the people who have declared their intent to migrate their keys and are waiting for the next mutation session.

Keys

The current individuals we recognise, but not necessarily yet included in a ring.

NextPersonalId

The next free and never reserved personal ID.

OnboardingQueue

Paginated collection of people public keys ready to be included in a ring.

OnboardingSize

Maximum number of people queued before onboarding to a ring.

PendingSuspensions

A map of all rings which currently have pending suspensions and need cleaning, along with their respective number of suspended keys which need to be removed.

People

The current individuals we recognise, but not necessarily yet included in a ring.

QueuePageIndices

Keeps track of the page indices of the head and tail of the onboarding queue.

ReservedPersonalId

Candidates’ reserved identities which we track.

RingBuildingPeopleLimit

Hint for the maximum number of people that can be included in a ring through a single root building call. If no value is set, then the onboarding size will be used instead.

RingKeys

Both the keys that are included in built rings and the keys that will be used in future rings.

RingKeysStatus

Stores the meta information for each ring, the number of keys and how many are actually included in the root.

RingsState

The state of the pallet regarding the actions that are currently allowed to be performed on all existing rings.

Root

The current individuals we recognise.